chore: separate Lesson views into their own files
