+++ iOS/PadawanWallet/Features/Lessons/LessonsRootView.swift
@@ -133,90 +133,3 @@ struct LessonsRootView: View {
         }
     }
 }
-
-struct LessonRow: View {
-    let lesson: Lesson
-    @Environment(\.padawanColors) private var colors
-
-    var body: some View {
-        ZStack {
-            // Shadow background
-            RoundedRectangle(cornerRadius: 12)
-                .fill(.black)
-                .offset(x: 4, y: 4)
-            
-            // Row content
-            HStack {
-                Text(lesson.title)
-                    .foregroundColor(colors.text)
-                    .padding()
-                Spacer()
-                if lesson.isHighlighted {
-                    Image(systemName: "star.fill")
-                        .foregroundColor(colors.accent2)
-                        .padding(.trailing, 12)
-                } else {
-                    Circle()
-                        .stroke(colors.textLight, lineWidth: 2)
-                        .frame(width: 24, height: 24)
-                        .padding(.trailing, 12)
-                }
-            }
-            .background(lesson.isHighlighted ? colors.accent1Light : colors.background2)
-            .cornerRadius(12)
-            .overlay(
-                RoundedRectangle(cornerRadius: 12)
-                    .stroke(.black, lineWidth: 2)
-            )
-        }
-    }
-}
-
-struct LessonDetailScreen: View {
-    let lesson: Lesson
-    @Environment(\.padawanColors) private var colors
-
-    var body: some View {
-        ScrollView {
-            VStack(alignment: .leading, spacing: 16) {
-                if let content = lesson.content {
-                    buildLessonContent(content: content)
-                } else {
-                    Text("Lesson content coming soon!")
-                        .font(.title2)
-                        .foregroundColor(colors.textLight)
-                        .padding()
-                }
-            }
-            .padding()
-        }
-        .background(colors.background)
-        .navigationTitle(lesson.title)
-        .navigationBarTitleDisplayMode(.inline)
-    }
-    
-    @ViewBuilder
-    private func buildLessonContent(content: LessonContent) -> some View {
-        ForEach(0..<content.elements.count, id: \.self) { index in
-            let element = content.elements[index]
-            
-            VStack(alignment: .leading, spacing: 12) {
-                Text(NSLocalizedString(element.resourceKey, comment: ""))
-                    .font(fontForElementType(element.type))
-                    .fontWeight(element.type == .title || element.type == .subtitle ? .bold : .regular)
-                    .lineSpacing(element.type == .paragraph ? 4 : 0)
-                    .padding(.top, element.type == .subtitle ? 8 : 0)
-                    .foregroundColor(element.type == .title ? colors.text : (element.type == .subtitle ? colors.text : colors.textLight))
-            }
-        }
-    }
-    
-    private func fontForElementType(_ type: ElementType) -> Font {
-        switch type {
-        case .title: return .largeTitle
-        case .subtitle: return .title2
-        case .paragraph: return .body
-        case .resource: return .body
-        }
-    }
-}
